Actualizacion  Spring 3
Se crearon las clases de los modulos Proveedores, Ventas, Detalle de Ventas, Clientes y Productos

+++ Proyecto_Ciclo_3/ciclo3back/src/main/java/co/edu/unbosque/ciclo3back/dao/ClientesDAO.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.ciclo3back.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.unbosque.ciclo3back.model.Clientes;
 
-public interface ClientesDAO extends JpaRepository<Clientes, Integer> {
+public interface ClientesDAO extends JpaRepository<Clientes, Long> {
 
 }
